refactor: split tests for asset manager into separate files
